fix(cucumber): fix indexer tests
