+++ integration_tests/tests/steps/indexer.rs
@@ -1,11 +1,18 @@
 //   Copyright 2023 The Tari Project
 //   SPDX-License-Identifier: BSD-3-Clause
 
-use std::str::FromStr;
+use std::{
+    collections::{BTreeMap, HashMap},
+    str::FromStr,
+};
 
-use cucumber::{then, when};
-use integration_tests::TariWorld;
+use cucumber::{given, then, when};
+use integration_tests::{
+    indexer::{spawn_indexer, IndexerProcess},
+    TariWorld,
+};
 use tari_comms::multiaddr::Multiaddr;
+use tari_crypto::tari_utilities::hex::Hex;
 use tari_indexer_client::types::AddPeerRequest;
 
 #[when(expr = "indexer {word} connects to all other validators")]
@@ -44,4 +51,133 @@ async fn indexer_has_scanned_to_height(world: &mut TariWorld, name: String, bloc
 
     let stats = client.get_epoch_manager_stats().await.expect("Failed to get stats");
     assert_eq!(stats.current_block_height, block_height);
+}
+
+#[given(expr = "an indexer {word} connected to base node {word}")]
+async fn start_indexer(world: &mut TariWorld, indexer_name: String, bn_name: String) {
+    spawn_indexer(world, indexer_name, bn_name).await;
+}
+
+#[given(expr = "{word} indexer GraphQL request works")]
+async fn works_indexer_graphql(world: &mut TariWorld, indexer_name: String) {
+    let indexer: &mut IndexerProcess = world.indexers.get_mut(&indexer_name).unwrap();
+    // insert event mock data in the substate manager database
+    indexer.insert_event_mock_data().await;
+    let mut graphql_client = indexer.get_graphql_indexer_client().await;
+    let component_address = [0u8; 32];
+    let template_address = [0u8; 32];
+    let tx_hash = [0u8; 32];
+    let query = format!(
+        "{{ getEventsForTransaction(txHash: {:?}) {{ componentAddress, templateAddress, txHash, topic, payload }}
+    }}",
+        tx_hash.to_hex()
+    );
+    let res = graphql_client
+        .send_request::<HashMap<String, Vec<tari_indexer::graphql::model::events::Event>>>(&query, None, None)
+        .await
+        .expect("Failed to obtain getEventsForTransaction query result");
+    let res = res.get("getEventsForTransaction").unwrap();
+    assert_eq!(res.len(), 1);
+    assert_eq!(res[0].component_address, Some(component_address));
+    assert_eq!(res[0].template_address, template_address);
+    assert_eq!(res[0].tx_hash, tx_hash);
+    assert_eq!(res[0].topic, "my_event");
+    assert_eq!(
+        res[0].payload,
+        BTreeMap::from([("my".to_string(), "event".to_string())])
+    );
+}
+
+#[when(expr = "indexer {word} scans the network {int} events for account {word} with topics {word}")]
+async fn indexer_scans_network_events(
+    world: &mut TariWorld,
+    indexer_name: String,
+    num_events: u32,
+    account_name: String,
+    topics: String,
+) {
+    let indexer: &mut IndexerProcess = world.indexers.get_mut(&indexer_name).unwrap();
+    let accounts_component_addresses = world.outputs.get(&account_name).expect("Account name not found");
+    let component_address = accounts_component_addresses
+        .into_iter()
+        .find(|(k, _)| k.contains("components/Account"))
+        .map(|(_, v)| v)
+        .expect("Did not find component address");
+
+    let mut graphql_client = indexer.get_graphql_indexer_client().await;
+    let query = format!(
+        r#"{{ getEventsForComponent(componentAddress: "{}", version: {}) {{ componentAddress, templateAddress, txHash, topic, payload }} }}"#,
+        component_address.address, component_address.version
+    );
+    let res = graphql_client
+        .send_request::<HashMap<String, Vec<tari_indexer::graphql::model::events::Event>>>(&query, None, None)
+        .await
+        .expect("Failed to obtain getEventsForComponent query result");
+
+    let events_for_component = res.get("getEventsForComponent").unwrap();
+    assert_eq!(
+        events_for_component.len(),
+        num_events as usize,
+        "Unexpected number of events returned got {}, expected {}",
+        events_for_component.len(),
+        num_events
+    );
+
+    let topics = topics.split(',').collect::<Vec<_>>();
+    assert_eq!(
+        topics.len(),
+        num_events as usize,
+        "Unexpected number of topics provided got {}, expected {}",
+        topics.len(),
+        num_events
+    );
+
+    for (ind, topic) in topics.iter().enumerate() {
+        let event = events_for_component[ind].clone();
+        assert_eq!(&event.topic, topic);
+    }
+}
+
+#[when(expr = "the indexer {word} tracks the address {word}")]
+async fn track_addresss_in_indexer(world: &mut TariWorld, indexer_name: String, output_ref: String) {
+    let indexer = world.indexers.get(&indexer_name).unwrap();
+    assert!(!indexer.handle.is_finished(), "Indexer {} is not running", indexer_name);
+    indexer.add_address(world, output_ref).await;
+}
+
+#[then(expr = "the indexer {word} returns version {int} for substate {word}")]
+async fn assert_indexer_substate_version(
+    world: &mut TariWorld,
+    indexer_name: String,
+    version: u32,
+    output_ref: String,
+) {
+    let indexer = world.indexers.get(&indexer_name).unwrap();
+    assert!(!indexer.handle.is_finished(), "Indexer {} is not running", indexer_name);
+    let substate = indexer.get_substate(world, output_ref, version).await;
+    eprintln!(
+        "indexer.get_substate result: {}",
+        serde_json::to_string_pretty(&substate).unwrap()
+    );
+    assert_eq!(substate.version, version);
+}
+
+#[then(expr = "the indexer {word} returns {int} non fungibles for resource {word}")]
+async fn assert_indexer_non_fungible_list(
+    world: &mut TariWorld,
+    indexer_name: String,
+    count: usize,
+    output_ref: String,
+) {
+    let indexer = world.indexers.get(&indexer_name).unwrap();
+    assert!(!indexer.handle.is_finished(), "Indexer {} is not running", indexer_name);
+    let nfts = indexer.get_non_fungibles(world, output_ref, 0, count as u64).await;
+    eprintln!("indexer.get_non_fungibles result: {:?}", nfts);
+    assert_eq!(
+        nfts.len(),
+        count,
+        "Unexpected number of NFTs returned. Expected: {}, Actual: {}",
+        count,
+        nfts.len()
+    );
 }
